--- app/javascript/common/schemaHelpers.js
@@ -1,3 +1,0 @@
-export const logSchemaError = err => {
-  console.error(err.name, err.errors); // eslint-disable-line no-console
-};
