+++ app/javascript/data/models/datastores.js
@@ -1,0 +1,10 @@
+import { array, string, object } from 'yup';
+
+export const datastoresSchema = array()
+  .of(
+    object().shape({
+      href: string().required(),
+      id: string().required()
+    })
+  )
+  .nullable();
