+++ app/javascript/data/models/provider.js
@@ -1,0 +1,11 @@
+import { array, string, object } from 'yup';
+
+export const providersSchema = array()
+  .of(
+    object().shape({
+      href: string().required(),
+      type: string().required(),
+      id: string().required()
+    })
+  )
+  .nullable();
